lint: reformat everything

+++ pkgs/stdenv/linux/make-bootstrap-tools.nix
@@ -1,13 +1,14 @@
-{ pkgs ? import ../../.. {} }:
+{ pkgs ? import ../../.. { } }:
 
 let
   libc = pkgs.stdenv.cc.libc;
-  patchelf = pkgs.patchelf.overrideAttrs(previousAttrs: {
-    NIX_CFLAGS_COMPILE = (previousAttrs.NIX_CFLAGS_COMPILE or []) ++ [ "-static-libgcc" "-static-libstdc++" ];
-    NIX_CFLAGS_LINK = (previousAttrs.NIX_CFLAGS_LINK or []) ++ [ "-static-libgcc" "-static-libstdc++" ];
+  patchelf = pkgs.patchelf.overrideAttrs (previousAttrs: {
+    NIX_CFLAGS_COMPILE = (previousAttrs.NIX_CFLAGS_COMPILE or [ ])
+      ++ [ "-static-libgcc" "-static-libstdc++" ];
+    NIX_CFLAGS_LINK = (previousAttrs.NIX_CFLAGS_LINK or [ ])
+      ++ [ "-static-libgcc" "-static-libstdc++" ];
   });
 in with pkgs; rec {
-
 
   coreutilsMinimal = coreutils.override (args: {
     # We want coreutils without ACL/attr support.
@@ -43,27 +44,26 @@ in with pkgs; rec {
     enableShared = false;
   };
 
-  build =
-    let
-      # ${libc.src}/sysdeps/unix/sysv/linux/loongarch/lp64/libnsl.abilist does not exist!
-      withLibnsl = !stdenv.hostPlatform.isLoongArch64;
-    in
-    stdenv.mkDerivation {
-      name = "stdenv-bootstrap-tools";
+  build = let
+    # ${libc.src}/sysdeps/unix/sysv/linux/loongarch/lp64/libnsl.abilist does not exist!
+    withLibnsl = !stdenv.hostPlatform.isLoongArch64;
+  in stdenv.mkDerivation {
+    name = "stdenv-bootstrap-tools";
 
-      meta = {
-        # Increase priority to unblock nixpkgs-unstable
-        # https://github.com/NixOS/nixpkgs/pull/104679#issuecomment-732267288
-        schedulingPriority = 200;
-      };
+    meta = {
+      # Increase priority to unblock nixpkgs-unstable
+      # https://github.com/NixOS/nixpkgs/pull/104679#issuecomment-732267288
+      schedulingPriority = 200;
+    };
 
-      nativeBuildInputs = [ buildPackages.nukeReferences buildPackages.cpio ];
+    nativeBuildInputs = [ buildPackages.nukeReferences buildPackages.cpio ];
 
-      buildCommand = ''
-        set -x
-        mkdir -p $out/bin $out/lib $out/libexec
+    buildCommand = ''
+      set -x
+      mkdir -p $out/bin $out/lib $out/libexec
 
-      '' + (if (stdenv.hostPlatform.libc == "glibc") then ''
+    '' + (if (stdenv.hostPlatform.libc == "glibc") then
+      ''
         # Copy what we need of Glibc.
         cp -d ${libc.out}/lib/ld*.so* $out/lib
         cp -d ${libc.out}/lib/libc*.so* $out/lib
@@ -100,18 +100,19 @@ in with pkgs; rec {
         find $out/include -name .install -exec rm {} \;
         find $out/include -name ..install.cmd -exec rm {} \;
         mv $out/include $out/include-glibc
-    '' else if (stdenv.hostPlatform.libc == "musl") then ''
-        # Copy what we need from musl
-        cp ${libc.out}/lib/* $out/lib
-        cp -rL ${libc.dev}/include $out
-        chmod -R u+w "$out"
+      ''
+    else if (stdenv.hostPlatform.libc == "musl") then ''
+      # Copy what we need from musl
+      cp ${libc.out}/lib/* $out/lib
+      cp -rL ${libc.dev}/include $out
+      chmod -R u+w "$out"
 
-        rm -rf $out/include/mtd $out/include/rdma $out/include/sound $out/include/video
-        find $out/include -name .install -exec rm {} \;
-        find $out/include -name ..install.cmd -exec rm {} \;
-        mv $out/include $out/include-libc
-    '' else throw "unsupported libc for bootstrap tools")
-    + ''
+      rm -rf $out/include/mtd $out/include/rdma $out/include/sound $out/include/video
+      find $out/include -name .install -exec rm {} \;
+      find $out/include -name ..install.cmd -exec rm {} \;
+      mv $out/include $out/include-libc
+    '' else
+      throw "unsupported libc for bootstrap tools") + ''
         # Copy coreutils, bash, etc.
         cp -d ${coreutilsMinimal.out}/bin/* $out/bin
         (cd $out/bin && rm vdir dir sha*sum pinky factor pathchk runcon shuf who whoami shred users)
@@ -206,44 +207,43 @@ in with pkgs; rec {
         nuke-refs $out/on-server/busybox
       ''; # */
 
-      # The result should not contain any references (store paths) so
-      # that we can safely copy them out of the store and to other
-      # locations in the store.
-      allowedReferences = [];
-    };
+    # The result should not contain any references (store paths) so
+    # that we can safely copy them out of the store and to other
+    # locations in the store.
+    allowedReferences = [ ];
+  };
 
   bootstrapFiles = {
     # Make them their own store paths to test that busybox still works when the binary is named /nix/store/HASH-busybox
-    busybox = runCommand "busybox" {} "cp ${build}/on-server/busybox $out";
-    bootstrapTools = runCommand "bootstrap-tools.tar.xz" {} "cp ${build}/on-server/bootstrap-tools.tar.xz $out";
+    busybox = runCommand "busybox" { } "cp ${build}/on-server/busybox $out";
+    bootstrapTools = runCommand "bootstrap-tools.tar.xz" { }
+      "cp ${build}/on-server/bootstrap-tools.tar.xz $out";
   };
 
-  bootstrapTools =
-    let extraAttrs = lib.optionalAttrs
-      config.contentAddressedByDefault
-      {
-        __contentAddressed = true;
-        outputHashAlgo = "sha256";
-        outputHashMode = "recursive";
-      };
-    in
-    if (stdenv.hostPlatform.libc == "glibc") then
+  bootstrapTools = let
+    extraAttrs = lib.optionalAttrs config.contentAddressedByDefault {
+      __contentAddressed = true;
+      outputHashAlgo = "sha256";
+      outputHashMode = "recursive";
+    };
+  in if (stdenv.hostPlatform.libc == "glibc") then
     import ./bootstrap-tools {
       inherit (stdenv.buildPlatform) system; # Used to determine where to build
       inherit bootstrapFiles extraAttrs;
     }
-    else if (stdenv.hostPlatform.libc == "musl") then
+  else if (stdenv.hostPlatform.libc == "musl") then
     import ./bootstrap-tools-musl {
       inherit (stdenv.buildPlatform) system; # Used to determine where to build
       inherit bootstrapFiles extraAttrs;
     }
-    else throw "unsupported libc";
+  else
+    throw "unsupported libc";
 
   test = derivation {
     name = "test-bootstrap-tools";
     inherit (stdenv.hostPlatform) system; # We cannot "cross test"
     builder = bootstrapFiles.busybox;
-    args = [ "ash" "-e" "-c" "eval \"$buildCommand\"" ];
+    args = [ "ash" "-e" "-c" ''eval "$buildCommand"'' ];
 
     buildCommand = ''
       export PATH=${bootstrapTools}/bin
@@ -261,7 +261,10 @@ in with pkgs; rec {
       gcc --version
 
     '' + lib.optionalString (stdenv.hostPlatform.libc == "glibc") ''
-      rtld=$(echo ${bootstrapTools}/lib/${builtins.unsafeDiscardStringContext /* only basename */ (builtins.baseNameOf binutils.dynamicLinker)})
+      rtld=$(echo ${bootstrapTools}/lib/${
+        builtins.unsafeDiscardStringContext # only basename
+        (builtins.baseNameOf binutils.dynamicLinker)
+      })
       libc_includes=${bootstrapTools}/include-glibc
     '' + lib.optionalString (stdenv.hostPlatform.libc == "musl") ''
       rtld=$(echo ${bootstrapTools}/lib/ld-musl*.so.?)
